+++ build.ps1
@@ -1,12 +1,23 @@
-﻿param(
+#Requires -Version 5.1
 
-    [string] $ikvmc = "c:\tools\ikvm-8.1.5717.0\bin\ikvmc.exe",
+param(
+
     [string] $version = "2.3.5",
     [string] $assemblyversion = "2.0.0",
     [string] $pre = $null
 )
 
+$ErrorActionPreference = 'Stop'
+$ProgressPreference = 'SilentlyContinue'
+
 New-Item bin -ItemType Directory -ErrorAction SilentlyContinue | Out-Null
+
+$baseUri = "https://repository.ow2.org/nexus/service/local/repositories/releases/content/org/ow2/sat4j"
+
+
+Write-Output "Downloading jars" | Out-Host
+Invoke-WebRequest -URI "$baseUri/org.ow2.sat4j.core/$version/org.ow2.sat4j.core-$version.jar" -OutFile bin/org.ow2.sat4j.core.jar
+Invoke-WebRequest -URI "$baseUri/org.ow2.sat4j.pb/$version/org.ow2.sat4j.pb-$version.jar" -OutFile bin/org.ow2.sat4j.core.jar
 
 if($pre) {
     $version += "-" + $pre
@@ -18,14 +29,14 @@ $ikvm_args = @(
     "-keyfile:..\featureide.snk",
     "-version:$assemblyversion",
     "-fileversion:$version",
-    "{", , "..\lib\org.sat4j.core.jar", "}",
-    "{", "..\lib\org.sat4j.pb.jar", "}"
+    "{", , "..\bin\org.sat4j.core.jar", "}",
+    "{", "..\bin\org.sat4j.pb.jar", "}"
 )
 
 Write-Output "Compiling jars" | Out-Host
 
 Push-Location bin
-& $ikvmc $ikvm_args
+ikvmc $ikvm_args
 Pop-Location
 
 Write-Output "Packing files" | Out-Host